bootinfo_load: recognize ASCII EOT as end of bootinfo
Mac OS 9's CHRP boot info consists of the standard SGML document terminated by
an ASCII EOT, followed by a large binary blob which is relocated by the Forth
script. This patch allows the bootinfo loader to stop loading the script if it
detects an EOT, avoiding the out-of-memory error previously encountered when
booting OS 9.

Signed-off-by: Cormac O'Brien <i.am.cormac.obrien@gmail.com>